#520 - simpler, efficient, more secure query
With the view updated with the additional columns, the query can return to the simpler, original version.  It turns out that the `= ANY()` technique is recommended over using the `IN` operator, for efficiency and other reasons.  Finally, the DB module can ensure the parameters are safe; parameterized queries prevent SQL injection vulnerabilities.